Revert "test: expand coverage for untested pure functions and handler paths (…"
This reverts commit 97699d7bed48283c0d25f413bbd1391f975ad7dd.
