+++ web/src/i18n/locales/en.json
@@ -25,7 +25,7 @@
 		"clear": "Clear",
 		"interrupted": "Interrupted",
 		"selectAll": "Select all",
-		"saving": "Saving…",
+		"saving": "Saving...",
 		"add": "Add",
 		"edit": "Edit",
 		"save": "Save",
@@ -34,13 +34,13 @@
 		"syncing": "Syncing…",
 		"refresh": "Refresh",
 		"refreshing": "Refreshing…",
-		"discovering": "Discovering…",
+		"discovering": "Discovering...",
 		"n_a": "N/A",
 		"current": "Current",
 		"optional": "Optional",
 		"required": "Required",
 		"done": "Done",
-		"creating": "Creating…",
+		"creating": "Creating...",
 		"resolving": "Resolving…",
 		"live": "Live",
 		"deleted": "Deleted",
@@ -383,7 +383,7 @@
 			"title": "Controls",
 			"stop": "Stop",
 			"generating": "Models are generating - click Stop to cancel",
-			"sendMessage": "Send a message…",
+			"sendMessage": "Send a message...",
 			"send": "Send"
 		},
 		"chatWithAi": "Chat with AI",
@@ -639,8 +639,8 @@
 		"btn_discover_all": "Discover All Models",
 		"btn_refresh_quotas": "Refresh Quotas/Balances",
 		"btn_add_provider": "Add Provider",
-		"btn_discovering": "Discovering…",
-		"btn_refreshing": "Refreshing…",
+		"btn_discovering": "Discovering...",
+		"btn_refreshing": "Refreshing...",
 		"filter_placeholder": "Filter providers…",
 		"filter_provider_type": "Provider type",
 		"filter_all": "All ({{count}})",
@@ -722,7 +722,7 @@
 		"card_last_discovery": "Last Discovery",
 		"card_btn_edit": "Edit",
 		"card_btn_discover": "Discover Models",
-		"card_btn_discovering": "Discovering…",
+		"card_btn_discovering": "Discovering...",
 		"card_btn_delete": "Delete",
 		"card_copy_name": "Click to copy provider name",
 		"card_copy_url": "Click to copy API base URL",
@@ -1025,7 +1025,7 @@
 			},
 			"noProvidersConfigured": "No providers configured.",
 			"discardChanges": "Discard unsaved changes?",
-			"loadingProviders": "Loading providers…",
+			"loadingProviders": "Loading providers...",
 			"stripReasoningDescriptionText": "When enabled, reasoning/thinking tokens are removed from streaming responses for clients that cannot handle them.",
 			"providerInstructionsText": "Click a provider to restrict access. All are accessible by default.",
 			"nameLabel": "Name",
@@ -1395,7 +1395,7 @@
 			"staleRequestTimeout.description": "Mark pending/streaming requests as \"interrupted\" if they remain in-progress longer than this. 0 to disable. Accounts for providers with long time-to-first-token.",
 			"deleteRequests": "Delete Requests",
 			"deleteRequests.tooltip": "Delete request logs older than a selected range",
-			"deleteRequests.selectRange": "Select range…",
+			"deleteRequests.selectRange": "Select range...",
 			"deleteRequests.olderThan1d": "Older than 1 day",
 			"deleteRequests.olderThan1w": "Older than 1 week",
 			"deleteRequests.olderThan1m": "Older than 1 month",
